comment in Authenticatedsessioncontroller

+++ routes/web.php
@@ -43,31 +43,31 @@ Route::middleware('auth')->group(function () {
 });
 
 
-Route::get('/Categories',[CategoryController::class, 'view'])->name('admin.home');
-Route::post('/Categories',[CategoryController::class, 'create'])->name('addCategory');
-Route::get('/statistics',[AdminController::class, 'statistics'])->name('statistics');
-Route::put('/Category',[CategoryController::class, 'update'])->name('updateCategory');
-Route::delete('/Categories/{category}',[CategoryController::class, 'delete'])->name('deleteCategory');
+Route::get('/Categories',[CategoryController::class, 'view'])->middleware(['auth', 'admin'])->name('admin.home');
+Route::post('/Categories',[CategoryController::class, 'create'])->middleware(['auth', 'admin'])->name('addCategory');
+Route::get('/statistics',[AdminController::class, 'statistics'])->middleware(['auth', 'admin'])->name('statistics');
+Route::put('/Category',[CategoryController::class, 'update'])->middleware(['auth', 'admin'])->name('updateCategory');
+Route::delete('/Categories/{category}',[CategoryController::class, 'delete'])->middleware(['auth', 'admin'])->name('deleteCategory');
 
-Route::get('/allEvents',[EventController::class, 'viewAll'])->name('allevents');
+Route::get('/allEvents',[EventController::class, 'viewAll'])->middleware(['auth', 'admin'])->name('allevents');
 
-Route::get('/Events',[EventController::class, 'view'])->name('Events');
-Route::post('/Events',[EventController::class, 'create'])->name('addEvent');
+Route::get('/Events',[EventController::class, 'view'])->middleware(['auth', 'organizer'])->name('Events');
+Route::post('/Events',[EventController::class, 'create'])->middleware(['auth', 'organizer'])->name('addEvent');
 Route::patch('/update-status/{event}', [EventController::class, 'updateStatus'])->name('updateStatus');
 
-Route::delete('/Events/{event}', [EventController::class, 'delete'])->name('deleteEvent');
-Route::put('/update-event', [EventController::class, 'update'])->name('updateEvent');
+Route::delete('/Events/{event}', [EventController::class, 'delete'])->middleware(['auth', 'organizer'])->name('deleteEvent');
+Route::put('/update-event', [EventController::class, 'update'])->middleware(['auth', 'organizer'])->name('updateEvent');
 Route::get('/moreDetails/{id}', [EventController::class, 'showDetails'])->name('moreDetails');
-Route::post('/create-reservation/{eventId}', [ReservationController::class, 'createReservation'])->name('createReservation');
-Route::get('/view-reservations/{eventId}', [ReservationController::class, 'viewReservations'])->name('viewReservations');
-Route::patch('/update-reservation-statut/{reservationId}', [ReservationController::class, 'updateReservationStatus'])->name('updateReservationStatus');
+Route::post('/create-reservation/{eventId}', [ReservationController::class, 'createReservation'])->middleware(['auth', 'client'])->name('createReservation');
+Route::get('/view-reservations/{eventId}', [ReservationController::class, 'viewReservations'])->middleware(['auth', 'organizer'])->name('viewReservations');
+Route::patch('/update-reservation-statut/{reservationId}', [ReservationController::class, 'updateReservationStatus'])->middleware(['auth', 'organizer'])->name('updateReservationStatus');
 Route::get('/generate-ticket/{reservation}/{event}', [ReservationController::class, 'generateTicket'])->name('generateTicket');
-Route::post('/events/search', [EventController::class, 'viewClient'])->name('events.search');
+Route::post('/events/search', [EventController::class, 'viewClient'])->middleware(['auth', 'client'])->name('events.search');
 
 
-Route::get('/Users', [AdminController::class, 'viewUsers'])->name('users');
-Route::put('/ban/user/{userId}',  [AdminController::class, 'banUser'])->name('ban.user');
-Route::put('/Unban/user/{userId}',  [AdminController::class, 'unbanUser'])->name('unban.user');
+Route::get('/Users', [AdminController::class, 'viewUsers'])->middleware(['auth', 'admin'])->name('users');
+Route::put('/ban/user/{userId}',  [AdminController::class, 'banUser'])->middleware(['auth', 'admin'])->name('ban.user');
+Route::put('/Unban/user/{userId}',  [AdminController::class, 'unbanUser'])->middleware(['auth', 'admin'])->name('unban.user');
